+++ src/afl-fuzz.c
@@ -76,7 +76,9 @@ static void at_exit() {
 
   }
 
-  u32 kill_signal = SIGKILL;
+  u8 kill_signal = SIGKILL;
+
+  /* AFL_KILL_SIGNAL should already be initialized by afl_fsrv_init() */
   if (getenv("AFL_KILL_SIGNAL")) {
 
     kill_signal = atoi(getenv("AFL_KILL_SIGNAL"));
